requirements.txt
tempmail-python
requests
colorama
pystyle
beautifulsoup4
pywebview
asyncio
aiohttp
py7zr